progress 애니메이션 구현

+++ AbamDiary/Observable/GageModel.swift
@@ -9,8 +9,8 @@ import Foundation
 
 class GageModel {
     
-    var morningDiaryCount: Observable<Double> = Observable(0)
-    var nightDiaryCount: Observable<Double> = Observable(0)
+    var morningDiaryCount: Observable<Float> = Observable(0)
+    var nightDiaryCount: Observable<Float> = Observable(0)
 
     // 사용자가 중간에 백업하고 일기를 삭제하는 경우에 Count는 남아있어야함(고민)
     func saveDiarysCount(completion: @escaping () -> Void) {
